+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- method verifyMfaPasswordless redirect to GET passwordless/verify in orchestrated flow
+- method verifyMfaPasswordless does not redirect with an error when the verification code is invalid and the maximum trials limit has not been exceeded in orchestrated flow
 
 ## [1.40.0] - 2026-02-12